requirements.txt
psycopg2-binary==2.9.11
openpyxl==3.1.5
pandas==3.0.2
